Implemented collisions using the AABB class.
Each gameObject has an AABB which auto updates to the scale - Check to see whether a gameObject overlaps using AABB::Overlaps() :)

Added a pip image ready for adding pips

BUG: Vector2 + operator override should not set any of the attributes, must have gotten confused with the += operator.

TODO: Override the rest of the operators on the Vector2 class, such as / += -= ect. as wells as float, int, doubles and other Vector2s.

TODO: The Texture2D class does not currently render in units per pixels: Using only the scale to determine the size of the Texture, this needs fixing.